+++ apps/web/.next/server/server-reference-manifest.js
@@ -1,1 +1,1 @@
-self.__RSC_SERVER_MANIFEST="{\n  \"node\": {},\n  \"edge\": {},\n  \"encryptionKey\": \"Sst9rtYq1Q27aMBTYpAV/+tnyr0McpEWYmXmfvJJdNo=\"\n}"
+self.__RSC_SERVER_MANIFEST="{\n  \"node\": {},\n  \"edge\": {},\n  \"encryptionKey\": \"IobdjWtRhrrA1lnFSfmtAom6O1TT6+qsUtzWsaY+1gI=\"\n}"